+++ src/domain/City.java
@@ -1,0 +1,26 @@
+package src.domain;
+
+public class City {
+    private String name;
+    private int latitude;
+    private int longitude;
+
+    public City(String name, int latitude, int longitude) {
+        this.name = name;
+        this.latitude = latitude;
+        this.longitude = longitude;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getLatitude() {
+        return latitude;
+    }
+
+    public int getLongitude() {
+        return longitude;
+    }
+}
+
